requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
boto3>=1.21.32
python-dotenv>=0.19.2
